python_requirements.txt
aiohttp==3.8.4
aiokafka==0.8.0
aiosignal==1.3.1
alembic==1.9.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
distro==1.8.0
python-jose==3.3.0
passlib==1.7.4
fastapi==0.89.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
netaddr==0.8.0
pamqp==3.2.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.4
pyTelegramBotAPI==4.9.0
python-telegram-bot==20.0
redis==4.4.2
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
telebot==0.0.5
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
wrapt==1.14.1
yarl==1.8.2
yookassa==2.3.4
zipp==3.11.0
openai==0.27.4
kafka-python==2.0.2
requests-toolbelt==0.10.1
